[refactor] Load submodules using Lua tables
