+++ openwrt-openwisp-monitoring/files/lib/openwisp/interfaces.lua
