+++ openwrt-openwisp-monitoring/files/lib/openwisp/monitoring.lua
@@ -1,0 +1,10 @@
+local monitoring = {}
+
+monitoring.dhcp = require('dhcp')
+monitoring.interfaces = require('interfaces')
+monitoring.neighbors = require('neighbors')
+monitoring.resources = require('resources')
+monitoring.utils = require('monitoring_utils')
+monitoring.wifi = require('wifi')
+
+return monitoring
