+++ openwrt-openwisp-monitoring/files/lib/openwisp/monitoring_utils.lua
