+++ openwrt-openwisp-monitoring/files/lib/openwisp/resources.lua
