+++ openwrt-openwisp-monitoring/files/sbin/netjson-monitoring.lua
@@ -1,187 +1,18 @@
 #!/usr/bin/env lua
 -- retrieve monitoring information
 -- and return it as NetJSON Output
-io = require('io')
-ubus_lib = require('ubus')
-cjson = require('cjson')
-nixio = require('nixio')
-uci = require('uci')
-uci_cursor = uci.cursor()
+local cjson = require('cjson')
 
--- split function
-function split(str, pat)
-    local t = {}
-    local fpat = "(.-)" .. pat
-    local last_end = 1
-    local s, e, cap = str:find(fpat, 1)
-    while s do
-        if s ~= 1 or cap ~= "" then
-            table.insert(t, cap)
-        end
-        last_end = e + 1
-        s, e, cap = str:find(fpat, last_end)
-    end
-    if last_end <= #str then
-        cap = str:sub(last_end)
-        table.insert(t, cap)
-    end
-    return t
-end
-
-local function has_value(tab, val)
-    for index, value in ipairs(tab) do
-        if value == val then
-            return true
-        end
-    end
-    return false
-end
-
-local function starts_with(str, start)
-    return str:sub(1, #start) == start
-end
-
--- parse /proc/net/arp
-function parse_arp()
-    arp_info = {}
-    for line in io.lines('/proc/net/arp 2> /dev/null') do
-        if line:sub(1, 10) ~= 'IP address' then
-            ip, hw, flags, mac, mask, dev = line:match("(%S+)%s+(%S+)%s+(%S+)%s+(%S+)%s+(%S+)%s+(%S+)")
-            table.insert(arp_info, {
-                ip = ip,
-                mac = mac,
-                interface = dev,
-                state = ''
-            })
-        end
-    end
-    return arp_info
-end
-
-function get_ip_neigh_json()
-    arp_info = {}
-    output = io.popen('ip -json neigh 2> /dev/null'):read()
-    if output ~= nil and pcall(function () json_output = cjson.decode(output) end) then
-        json_output = cjson.decode(output)
-        for _, arp_entry in pairs(json_output) do
-            table.insert(arp_info, {
-                ip = arp_entry["dst"],
-                mac = arp_entry["lladdr"],
-                interface = arp_entry["dev"],
-                state = arp_entry["state"][1]
-            })
-        end
-    end
-    return arp_info
-end
-
-function get_ip_neigh()
-    arp_info = {}
-    output = io.popen('ip neigh 2> /dev/null')
-    for line in output:lines() do
-        ip, dev, mac, state = line:match("(%S+)%s+dev%s+(%S+)%s+lladdr%s+(%S+).*%s(%S+)")
-        if mac ~= nil then
-            table.insert(arp_info, {
-                ip = ip,
-                mac = mac,
-                interface = dev,
-                state = state
-            })
-        end
-    end
-    return arp_info
-end
-
-function get_neighbors()
-    arp_table = get_ip_neigh_json()
-    if next(arp_table) == nil then
-        arp_table = get_ip_neigh()
-    end
-    if next(arp_table) == nil then
-        arp_table = parse_arp()
-    end
-    return arp_table
-end
-
-function parse_dhcp_lease_file(path, leases)
-    local f = io.open(path, 'r')
-    if not f then
-        return leases
-    end
-
-    for line in f:lines() do
-        local expiry, mac, ip, name, id = line:match('(%S+)%s+(%S+)%s+(%S+)%s+(%S+)%s+(%S+)')
-        table.insert(leases, {
-            expiry = tonumber(expiry),
-            mac = mac,
-            ip = ip,
-            client_name = name,
-            client_id = id
-        })
-    end
-
-    return leases
-end
-
-function get_dhcp_leases()
-    local dhcp_configs = uci_cursor:get_all('dhcp')
-    local leases = {}
-
-    if not dhcp_configs or not next(dhcp_configs) then
-        return nil
-    end
-
-    for name, config in pairs(dhcp_configs) do
-        if config and config['.type'] == 'dnsmasq' and config.leasefile then
-            leases = parse_dhcp_lease_file(config.leasefile, leases)
-        end
-    end
-    return leases
-end
-
-function is_table_empty(table_)
-    return not table_ or next(table_) == nil
-end
-
-function parse_hostapd_clients(clients)
-  local data = {}
-  for mac, properties in pairs(clients) do
-      properties.mac = mac
-      table.insert(data, properties)
-  end
-  return data
-end
-
-function parse_iwinfo_clients(clients)
-  local data = {}
-  for i, p in pairs(clients) do
-      client = {}
-      client.ht = p.rx.ht
-      client.mac = p.mac
-      client.authorized = p.authorized
-      client.vht = p.rx.vht
-      client.wmm = p.wme
-      client.mfp = p.mfp
-      client.auth = p.authenticated
-      client.signal = p.signal
-      client.noise = p.noise
-      table.insert(data, client)
-  end
-  return data
-end
-
--- takes ubus wireless.status clients output and converts it to NetJSON
-function netjson_clients(clients, is_mesh)
-    return (is_mesh and parse_iwinfo_clients(clients) or parse_hostapd_clients(clients))
-end
-
-ubus = ubus_lib.connect()
+local ubus_lib = require('ubus')
+local ubus = ubus_lib.connect()
 if not ubus then
     error('Failed to connect to ubusd')
 end
 
+local monitoring = require('openwisp.monitoring')
+
 -- helpers
-iwinfo_modes = {
+local iwinfo_modes = {
     ['Master'] = 'access_point',
     ['Client'] = 'station',
     ['Mesh Point'] = '802.11s',
@@ -189,63 +20,15 @@ iwinfo_modes = {
 }
 
 -- collect system info
-system_info = ubus:call('system', 'info', {})
-board = ubus:call('system', 'board', {})
-loadavg_output = io.popen('cat /proc/loadavg'):read()
-loadavg_output = split(loadavg_output, ' ')
-load_average = {tonumber(loadavg_output[1]), tonumber(loadavg_output[2]), tonumber(loadavg_output[3])}
+local system_info = ubus:call('system', 'info', {})
+local board = ubus:call('system', 'board', {})
+local loadavg_output = io.popen('cat /proc/loadavg'):read()
+loadavg_output = monitoring.utils.split(loadavg_output, ' ')
+local load_average = {tonumber(loadavg_output[1]), tonumber(loadavg_output[2]), tonumber(loadavg_output[3])}
 
-function parse_disk_usage()
-    file = io.popen('df')
-    disk_usage_info = {}
-    for line in file:lines() do
-        if line:sub(1, 10) ~= 'Filesystem' then
-            filesystem, size, used, available, percent, location =
-                line:match('(%S+)%s+(%S+)%s+(%S+)%s+(%S+)%s+(%S+)%s+(%S+)')
-            if filesystem ~= 'tmpfs' and not string.match(filesystem, 'overlayfs') then
-                percent = percent:gsub('%W', '')
-                -- available, size and used are in KiB
-                table.insert(disk_usage_info, {
-                    filesystem = filesystem,
-                    available_bytes = tonumber(available) * 1024,
-                    size_bytes = tonumber(size) * 1024,
-                    used_bytes = tonumber(used) * 1024,
-                    used_percent = tonumber(percent),
-                    mount_point = location
-                })
-            end
-        end
-    end
-    file:close()
-    return disk_usage_info
-end
-
-function get_cpus()
-    processors = io.popen('cat /proc/cpuinfo | grep -c processor')
-    cpus = tonumber(processors:read('*a'))
-    processors:close()
-    return cpus
-end
-
-function get_vpn_interfaces()
-    -- only openvpn supported for now
-    local items = uci_cursor:get_all('openvpn')
-    local vpn_interfaces = {}
-
-    if is_table_empty(items) then
-        return {}
-    end
-
-    for name, config in pairs(items) do
-        if config and config.dev then
-            vpn_interfaces[config.dev] = true
-        end
-    end
-    return vpn_interfaces
-end
 
 -- init netjson data structure
-netjson = {
+local netjson = {
     type = 'DeviceMonitoring',
     general = {
         hostname = board.hostname,
@@ -256,232 +39,53 @@ netjson = {
         load = load_average,
         memory = system_info.memory,
         swap = system_info.swap,
-        cpus = get_cpus(),
-        disk = parse_disk_usage()
+        cpus = monitoring.resources.get_cpus(),
+        disk = monitoring.resources.parse_disk_usage()
     }
 }
 
-dhcp_leases = get_dhcp_leases()
-if not is_table_empty(dhcp_leases) then
+local dhcp_leases = monitoring.dhcp.get_dhcp_leases()
+if not monitoring.utils.is_table_empty(dhcp_leases) then
     netjson.dhcp_leases = dhcp_leases
 end
 
-neighbors = get_neighbors()
-if not is_table_empty(neighbors) then
-    netjson.neighbors = neighbors
+local host_neighbors = monitoring.neighbors.get_neighbors()
+if not monitoring.utils.is_table_empty(host_neighbors) then
+    netjson.neighbors = host_neighbors
 end
 
 -- determine the interfaces to monitor
-traffic_monitored = arg[1]
-include_stats = {}
+local traffic_monitored = arg[1]
+local include_stats = {}
 if traffic_monitored and traffic_monitored ~= '*' then
-    traffic_monitored = split(traffic_monitored, ' ')
-    for i, name in pairs(traffic_monitored) do
+    traffic_monitored = monitoring.utils.split(traffic_monitored, ' ')
+    for _, name in pairs(traffic_monitored) do
         include_stats[name] = true
     end
 end
 
-function is_excluded(name)
-    return name == 'lo'
-end
-
-function find_default_gateway(routes)
-    for i = 1, #routes do
-        if routes[i].target == '0.0.0.0' then
-            return routes[i].nexthop
-        end
-    end
-    return nil
-end
-
 -- collect device data
-network_status = ubus:call('network.device', 'status', {})
-wireless_status = ubus:call('network.wireless', 'status', {})
-interface_data = ubus:call('network.interface', 'dump', {})
-nixio_data = nixio.getifaddrs()
-vpn_interfaces = get_vpn_interfaces()
-wireless_interfaces = {}
-interfaces = {}
-dns_servers = {}
-dns_search = {}
+local network_status = ubus:call('network.device', 'status', {})
+local wireless_status = ubus:call('network.wireless', 'status', {})
+local vpn_interfaces = monitoring.interfaces.get_vpn_interfaces()
+local wireless_interfaces = {}
+local host_interfaces = {}
+local dns_servers = {}
+local dns_search = {}
 
-function new_address_array(address, interface, family)
-    proto = interface['proto']
-    if proto == 'dhcpv6' then
-        proto = 'dhcp'
-    end
-    new_address = {
-        address = address['address'],
-        mask = address['mask'],
-        proto = proto,
-        family = family,
-        gateway = find_default_gateway(interface.route)
-    }
-    return new_address
-end
-
-specialized_interfaces = {
-    modemmanager = function(name, interface)
-        local modem = uci_cursor.get('network', interface['interface'], 'device')
-        local info = {}
-
-        local general = io.popen('mmcli --output-json -m '..modem):read("*a")
-        if general and pcall(function () general = cjson.decode(general) end) then
-            general = general.modem
-
-            if not is_table_empty(general['3gpp']) then
-                info.imei = general['3gpp'].imei
-                info.operator_name = general['3gpp']['operator-name']
-                info.operator_code = general['3gpp']['operator-code']
-            end
-
-            if not is_table_empty(general.generic) then
-                info.manufacturer = general.generic.manufacturer
-                info.model = general.generic.model
-                info.connection_status = general.generic.state
-                info.power_status = general.generic['power-state']
-            end
-        end
-
-        local signal = io.popen('mmcli --output-json -m '..modem..' --signal-get'):read()
-        if signal and pcall(function () signal = cjson.decode(signal) end) then
-            -- only send data if not empty to avoid generating too much traffic
-            if not is_table_empty(signal.modem) and not is_table_empty(signal.modem.signal) then
-                -- omit refresh rate
-                signal.modem.signal.refresh = nil
-                info.signal = {}
-                -- collect section and values only if not empty
-                for section_key, section_values in pairs(signal.modem.signal) do
-                    for key, value in pairs(section_values) do
-                        if value ~= '--' then
-                            if is_table_empty(info.signal[section_key]) then
-                                info.signal[section_key] = {}
-                            end
-                            info.signal[section_key][key] = tonumber(value)
-                        end
-                    end
-                end
-            end
-        end
-
-        return {type='modem-manager', mobile=info}
-    end
-}
-
-function get_interface_info(name, netjson_interface)
-    info = {
-        dns_search = nil,
-        dns_servers = nil
-    }
-    for _, interface in pairs(interface_data['interface']) do
-        if interface['l3_device'] == name then
-            if next(interface['dns-search']) then
-                info.dns_search = interface['dns-search']
-            end
-            if next(interface['dns-server']) then
-                info.dns_servers = interface['dns-server']
-            end
-            if netjson_interface.type == 'bridge' then
-                info.stp = uci_cursor.get('network', interface['interface'], 'stp') == '1'
-            end
-            -- collect specialized info if available
-            local specialized_info = specialized_interfaces[interface.proto]
-            if specialized_info then
-                info.specialized = specialized_info(name, interface)
-            end
-        end
-    end
-    return info
-end
-
-function array_concat(source, destination)
-    table.foreach(source, function(key, value)
-        table.insert(destination, value)
-    end)
-end
-
-function dict_merge(source, destination)
-    table.foreach(source, function(key, value)
-        destination[key] = value
-    end)
-end
-
--- collect interface addresses
-function get_addresses(name)
-    addresses = {}
-    interface_list = interface_data['interface']
-    addresses_list = {}
-    for _, interface in pairs(interface_list) do
-        if interface['l3_device'] == name then
-            proto = interface['proto']
-            if proto == 'dhcpv6' then
-                proto = 'dhcp'
-            end
-            for _, address in pairs(interface['ipv4-address']) do
-                table.insert(addresses_list, address['address'])
-                new_address = new_address_array(address, interface, 'ipv4')
-                table.insert(addresses, new_address)
-            end
-            for _, address in pairs(interface['ipv6-address']) do
-                table.insert(addresses_list, address['address'])
-                new_address = new_address_array(address, interface, 'ipv6')
-                table.insert(addresses, new_address)
-            end
-        end
-    end
-    for i = 1, #nixio_data do
-        if nixio_data[i].name == name then
-            if not is_excluded(name) then
-                family = nixio_data[i].family
-                addr = nixio_data[i].addr
-                if family == 'inet' then
-                    family = 'ipv4'
-                    -- Since we don't already know this from the dump, we can
-                    -- consider this dynamically assigned, this is the case for
-                    -- example for OpenVPN interfaces, which get their address
-                    -- from the DHCP server embedded in OpenVPN
-                    proto = 'dhcp'
-                elseif family == 'inet6' then
-                    family = 'ipv6'
-                    if starts_with(addr, 'fe80') then
-                        proto = 'static'
-                    else
-                        ula = uci_cursor.get('network', 'globals', 'ula_prefix')
-                        ula_prefix = split(ula, '::')[1]
-                        if starts_with(addr, ula_prefix) then
-                            proto = 'static'
-                        else
-                            proto = 'dhcp'
-                        end
-                    end
-                end
-                if family == 'ipv4' or family == 'ipv6' then
-                    if not has_value(addresses_list, addr) then
-                        table.insert(addresses, {
-                            address = addr,
-                            mask = nixio_data[i].prefix,
-                            proto = proto,
-                            family = family
-                        })
-                    end
-                end
-            end
-        end
-    end
-    return addresses
-end
 
 -- collect relevant wireless interface stats
 -- (traffic and connected clients)
-for radio_name, radio in pairs(wireless_status) do
-    for i, interface in ipairs(radio.interfaces) do
-        name = interface.ifname
+for _, radio in pairs(wireless_status) do
+    for _, interface in ipairs(radio.interfaces) do
+        local name = interface.ifname
         local is_mesh = false
-        if name and not is_excluded(name) then
-            iwinfo = ubus:call('iwinfo', 'info', {
+        local clients = nil
+        if name and not monitoring.utils.is_excluded(name) then
+            local iwinfo = ubus:call('iwinfo', 'info', {
                 device = name
             })
-            netjson_interface = {
+            local netjson_interface = {
                 name = name,
                 type = 'wireless',
                 wireless = {
@@ -501,24 +105,24 @@ for radio_name, radio in pairs(wireless_status) do
                 }).results
                 is_mesh = true
             else
-              hostapd_output = ubus:call('hostapd.' .. name, 'get_clients', {})
+              local hostapd_output = ubus:call('hostapd.' .. name, 'get_clients', {})
               if hostapd_output then
                   clients = hostapd_output.clients
               end
             end
-            if clients and next(clients) ~= nil then
-                netjson_interface.wireless.clients = netjson_clients(clients, is_mesh)
+            if not monitoring.utils.is_table_empty(clients) then
+                netjson_interface.wireless.clients = monitoring.wifi.netjson_clients(clients, is_mesh)
             end
             wireless_interfaces[name] = netjson_interface
         end
     end
 end
 
-function needs_inversion(interface)
+local function needs_inversion(interface)
     return interface.type == 'wireless' and interface.wireless.mode == 'access_point'
 end
 
-function invert_rx_tx(interface)
+local function invert_rx_tx(interface)
     for k, v in pairs(interface) do
         if string.sub(k, 0, 3) == "rx_" then
             local tx_key = "tx_" .. string.sub(k, 4)
@@ -533,8 +137,8 @@ end
 -- collect interface stats
 for name, interface in pairs(network_status) do
     -- only collect data from iterfaces which have not been excluded
-    if not is_excluded(name) then
-        netjson_interface = {
+    if not monitoring.utils.is_excluded(name) then
+        local netjson_interface = {
             name = name,
             type = string.lower(interface.type),
             up = interface.up,
@@ -546,11 +150,11 @@ for name, interface in pairs(network_status) do
             multicast = interface.multicast
         }
         if wireless_interfaces[name] then
-            dict_merge(wireless_interfaces[name], netjson_interface)
+            monitoring.utils.dict_merge(wireless_interfaces[name], netjson_interface)
             interface.type = netjson_interface.type
         end
         if interface.type == 'Network device' then
-            link_supported = interface['link-supported']
+            local link_supported = interface['link-supported']
             if link_supported and next(link_supported) then
                 netjson_interface.type = 'ethernet'
                 netjson_interface.link_supported = link_supported
@@ -569,11 +173,11 @@ for name, interface in pairs(network_status) do
             end
             netjson_interface.statistics = interface.statistics
         end
-        addresses = get_addresses(name)
+        local addresses = monitoring.interfaces.get_addresses(name)
         if next(addresses) then
             netjson_interface.addresses = addresses
         end
-        info = get_interface_info(name, netjson_interface)
+        local info = monitoring.interfaces.get_interface_info(name, netjson_interface)
         if info.stp ~= nil then
             netjson_interface.stp = info.stp
         end
@@ -582,19 +186,19 @@ for name, interface in pairs(network_status) do
                 netjson_interface[key] = value
             end
         end
-        table.insert(interfaces, netjson_interface)
+        table.insert(host_interfaces, netjson_interface)
         -- DNS info is independent from interface
         if info.dns_servers then
-            array_concat(info.dns_servers, dns_servers)
+            monitoring.utils.array_concat(info.dns_servers, dns_servers)
         end
         if info.dns_search then
-            array_concat(info.dns_search, dns_search)
+            monitoring.utils.array_concat(info.dns_search, dns_search)
         end
     end
 end
 
-if next(interfaces) ~= nil then
-    netjson.interfaces = interfaces
+if next(host_interfaces) ~= nil then
+    netjson.interfaces = host_interfaces
 end
 if next(dns_servers) ~= nil then
     netjson.dns_servers = dns_servers
